+++ sdk/storage/azure-storage-blobs/test/ut/storage_timeout_test.cpp
@@ -1,0 +1,72 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#include <chrono>
+#include <memory>
+#include <vector>
+
+#include <azure/storage/blobs.hpp>
+
+#include "test_base.hpp"
+
+namespace Azure { namespace Storage { namespace Test {
+
+  class PeekHttpRequestPolicy : public Core::Http::Policies::HttpPolicy {
+  public:
+    PeekHttpRequestPolicy(std::function<void(const Core::Http::Request&)> callback)
+        : m_callback(std::move(callback))
+    {
+    }
+
+    std::unique_ptr<Core::Http::RawResponse> Send(
+        Core::Http::Request& request,
+        Core::Http::Policies::NextHttpPolicy nextHttpPolicy,
+        Core::Context const& context) const override
+    {
+      m_callback(request);
+      return nextHttpPolicy.Send(request, context);
+    }
+
+    std::unique_ptr<HttpPolicy> Clone() const override
+    {
+      return std::make_unique<PeekHttpRequestPolicy>(*this);
+    }
+
+  private:
+    std::function<void(const Core::Http::Request&)> m_callback;
+  };
+
+  TEST(StoragetimeoutTest, Basic)
+  {
+    Azure::Nullable<int64_t> timeout;
+    auto callback = [&timeout](const Core::Http::Request& request) {
+      auto queryParameteres = request.GetUrl().GetQueryParameters();
+      auto ite = queryParameteres.find("timeout");
+      if (ite == queryParameteres.end())
+      {
+        timeout.Reset();
+      }
+      else
+      {
+        timeout = std::stoll(ite->second);
+      }
+    };
+
+    auto peekPolicyPtr = std::make_unique<PeekHttpRequestPolicy>(callback);
+    Blobs::BlobClientOptions clientOptions;
+    clientOptions.PerRetryPolicies.emplace_back(std::move(peekPolicyPtr));
+    auto containerClient = Azure::Storage::Blobs::BlobContainerClient::CreateFromConnectionString(
+        StandardStorageConnectionString(), LowercaseRandomString(), clientOptions);
+    containerClient.DeleteIfExists();
+    EXPECT_FALSE(timeout.HasValue());
+
+    Azure::Core::Context context;
+    context = context.WithDeadline(std::chrono::system_clock::now() + std::chrono::seconds(300));
+    containerClient.DeleteIfExists(Storage::Blobs::DeleteBlobContainerOptions(), context);
+
+    ASSERT_TRUE(timeout.HasValue());
+    EXPECT_GE(timeout.Value(), 299);
+    EXPECT_LE(timeout.Value(), 301);
+  }
+
+}}} // namespace Azure::Storage::Test