+++ sdk/storage/azure-storage-common/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 12.0.0-beta.10 (Unreleased)
 
+### New Features
+
+- Added server timeout support.
+
 ### Breaking Changes
 
 - Removed `Azure::Storage::Common::PackageVersion`.